Moving gce module to contrib.
Also fixing a broken import of oauth2client.appengine in
oauth2client.client. This reveals an untested code-path that
could have been found via 100% code coverage.
