+++ docs/source/oauth2client.contrib.gce.rst
@@ -1,0 +1,7 @@
+oauth2client.contrib.gce module
+===============================
+
+.. automodule:: oauth2client.contrib.gce
+    :members:
+    :undoc-members:
+    :show-inheritance:
