+++ docs/source/oauth2client.contrib.rst
@@ -16,6 +16,7 @@ Submodules
    oauth2client.contrib.appengine
    oauth2client.contrib.django_orm
    oauth2client.contrib.flask_util
+   oauth2client.contrib.gce
    oauth2client.contrib.keyring_storage
    oauth2client.contrib.locked_file
    oauth2client.contrib.multistore_file
