--- docs/source/oauth2client.gce.rst
@@ -1,7 +1,0 @@
-oauth2client.gce module
-=======================
-
-.. automodule:: oauth2client.gce
-    :members:
-    :undoc-members:
-    :show-inheritance:
