+++ oauth2client/contrib/gce.py
